requirements.txt
Flask==2.0.2
stripe==2.62.0
python-dotenv==0.19.2  # For loading environment variables from .env file
flask-cors==4.0.2  # For handling Cross-Origin Resource Sharing
